src/backend/requirements.txt
azure-functions==1.17.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
